+++ CHANGELOG.md
@@ -9,3 +9,4 @@ Changelog
 - Fix issue with failing to read in saved passwords (such as switching between computers)
 - Fix issues with old Java paths causing issues launching packs
 - Some changes to logging to the console to keep things in order under large loads
+- Add a way for semi public packs to auto add their codes to users launcher